+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - enhance regex alternations. ex: [01][0-9]|2[0-3] => ([01][0-9]|2[0-3])
 
 ### Fixed
+- Paste in inputmask #2088
 - The first character is truncated when masking. #2089
 - No leading zero for single-digit hours KO #2061
 - Only 1st placeholder is displayed for fields with same alias numeric #2060